reformat in the pattern
