+++ iogram_bot/__init__.py
@@ -1,0 +1,2 @@
+from . import handlers
+from . import config_data
