--- iogram_bot/base.py
@@ -1,34 +1,0 @@
-import logging
-import os
-from aiogram import Bot, Dispatcher
-from aiogram.filters import Command
-from aiogram.types import Message
-
-from dotenv import load_dotenv, find_dotenv
-if not find_dotenv():
-    exit('Переменные окружения не загружены т.к отсутствует файл .env')
-else:
-    load_dotenv()
-API_TOKEN: str = os.environ.get('BOT_TOKEN')
-
-
-# Создаем объекты бота и диспетчера
-bot: Bot = Bot(token=API_TOKEN)
-dp: Dispatcher = Dispatcher()
-
-
-# Этот хэндлер будет срабатывать на команду "/start"
-@dp.message(Command(commands=["start", 'help']))
-async def process_start_command(message: Message):
-    await message.answer('Привет!\nМеня зовут Эхо-бот!\nНапиши мне что-нибудь')
-
-
-# Этот хэндлер будет срабатывать на любые ваши текстовые сообщения,
-# кроме команд "/start" и "/help"
-@dp.message()
-async def send_echo(message: Message):
-    await message.reply(text=message.text)
-    print(message.json(indent=4, exclude_none=True))
-
-if __name__ == '__main__':
-    dp.run_polling(bot)
