+++ iogram_bot/config_data/__init__.py
@@ -1,0 +1,1 @@
+from . import config
