+++ main.py
@@ -1,0 +1,4 @@
+from iogram_bot.config_data.config import bot_dispatcher, bot
+
+if __name__ == '__main__':
+    bot_dispatcher.run_polling(bot)
